pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kg-agent-quant"
version = "0.1.0"
description = "Knowledge Graph Enhanced Alpha Factor Research with LLM Verification"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Research Team", email = "research@example.com"}
]
keywords = ["quantitative-finance", "alpha-factors", "knowledge-graph", "llm", "factor-investing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.5.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "pyyaml>=6.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.950",
]
qlib = [
    "pyqlib>=0.9.0",
    "lightgbm>=3.0.0",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.5.0",
]
viz = [
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "plotly>=5.0.0",
]
all = [
    "kg-agent-quant[qlib,llm,viz,dev]",
]

[project.urls]
Homepage = "https://github.com/YOUR_ORG/kg-agent-quant"
Documentation = "https://kg-agent-quant.readthedocs.io"
Repository = "https://github.com/YOUR_ORG/kg-agent-quant"
Issues = "https://github.com/YOUR_ORG/kg-agent-quant/issues"

[project.scripts]
kg-quant = "kg_quant.__main__:main"
kg-factor = "kg_quant.cli:main"
kg-eval = "kg_quant.evaluation.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
kg_quant = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"